+++ tests/basic_functionality.rs
@@ -1,7 +1,7 @@
 use message_io::events::{EventQueue};
-use message_io::network::{NetworkManager, NetEvent};
+use message_io::network::{NetworkManager, NetEvent, MAX_UDP_LEN};
 
-const MESSAGE_DATA: &'static str = "Small message";
+const SMALL_MESSAGE: &'static str = "Small message";
 
 #[test]
 fn simple_connection_data_disconnection_by_tcp() {
@@ -17,7 +17,7 @@ fn simple_connection_data_disconnection_by_tcp() {
             match event_queue.receive() {
                 NetEvent::Message(endpoint, message) => {
                     assert_eq!(*client_endpoint.as_ref().unwrap(), endpoint);
-                    assert_eq!(message, MESSAGE_DATA);
+                    assert_eq!(message, SMALL_MESSAGE);
                     network.send(endpoint, message).unwrap();
                 }
                 NetEvent::AddedEndpoint(endpoint) => {
@@ -39,12 +39,12 @@ fn simple_connection_data_disconnection_by_tcp() {
         let mut network = NetworkManager::new(move |net_event| sender.send(net_event));
 
         let server_endpoint = network.connect_tcp(server_addr).unwrap();
-        network.send(server_endpoint, MESSAGE_DATA.to_string()).unwrap();
+        network.send(server_endpoint, SMALL_MESSAGE.to_string()).unwrap();
         loop {
             match event_queue.receive() {
                 NetEvent::Message(endpoint, message) => {
                     assert_eq!(server_endpoint, endpoint);
-                    assert_eq!(message, MESSAGE_DATA);
+                    assert_eq!(message, SMALL_MESSAGE);
                     network.send(endpoint, message).unwrap();
                     break //Exit from thread, the connection will be automatically close
                 }
@@ -73,7 +73,7 @@ fn simple_data_by_udp() {
             match event_queue.receive() {
                 NetEvent::Message(endpoint, message) => {
                     assert_eq!(upd_listen_resource_id, endpoint.resource_id());
-                    assert_eq!(message, MESSAGE_DATA);
+                    assert_eq!(message, SMALL_MESSAGE);
                     network.send(endpoint, message).unwrap();
                     break //Exit from thread
                 }
@@ -88,12 +88,12 @@ fn simple_data_by_udp() {
         let mut network = NetworkManager::new(move |net_event| sender.send(net_event));
 
         let server_endpoint = network.connect_udp(server_addr).unwrap();
-        network.send(server_endpoint, MESSAGE_DATA.to_string()).unwrap();
+        network.send(server_endpoint, SMALL_MESSAGE.to_string()).unwrap();
         loop {
             match event_queue.receive() {
                 NetEvent::Message(endpoint, message) => {
                     assert_eq!(server_endpoint, endpoint);
-                    assert_eq!(message, MESSAGE_DATA);
+                    assert_eq!(message, SMALL_MESSAGE);
                     break //Exit from thread
                 }
                 _ => unreachable!(),
@@ -103,4 +103,78 @@ fn simple_data_by_udp() {
 
     server_handle.join().unwrap();
     client_handle.join().unwrap();
+}
+
+#[test]
+fn long_tcp_message() {
+    let mut event_queue = EventQueue::<NetEvent<Vec<u8>>>::new();
+    let sender = event_queue.sender().clone();
+    let mut network = NetworkManager::new(move |net_event| sender.send(net_event));
+
+    const MESSAGE_SIZE: usize = 1_000_000; // Arround 1MB
+    const VALUE: u8 = 0xAA;
+    let (_, receiver_addr) = network.listen_tcp("127.0.0.1:0").unwrap();
+
+    let receiver_handle = std::thread::spawn(move || {
+        // Pass the network to the thread. The network should be destroyed before event queue.
+        let _ = network;
+        loop {
+            match event_queue.receive() {
+                NetEvent::Message(_, message) => {
+                    assert_eq!(message.len(), MESSAGE_SIZE);
+                    assert!(message.iter().all(|&byte| byte == VALUE));
+                    break
+                }
+                NetEvent::AddedEndpoint(_) => (),
+                NetEvent::RemovedEndpoint(_) => (),
+            }
+        }
+    });
+
+    let mut event_queue = EventQueue::<NetEvent<Vec<u8>>>::new();
+    let sender = event_queue.sender().clone();
+    let mut network = NetworkManager::new(move |net_event| sender.send(net_event));
+
+    let receiver = network.connect_tcp(receiver_addr).unwrap();
+    let message = std::iter::repeat(VALUE).take(MESSAGE_SIZE).collect::<Vec<_>>();
+    network.send(receiver, message.clone()).unwrap(); // Blocks until the message is sent
+
+    receiver_handle.join().unwrap();
+}
+
+#[test]
+fn max_udp_size_message() {
+    let mut event_queue = EventQueue::<NetEvent<Vec<u8>>>::new();
+    let sender = event_queue.sender().clone();
+    let mut network = NetworkManager::new(move |net_event| sender.send(net_event));
+
+    const MESSAGE_SIZE: usize = MAX_UDP_LEN - (8 + 4); // Vec<u8> header + encoding header
+    const VALUE: u8 = 0xFF;
+    let (_, receiver_addr) = network.listen_udp("127.0.0.1:0").unwrap();
+
+    let receiver_handle = std::thread::spawn(move || {
+        // Pass the network to the thread. The network should be destroyed before event queue.
+        let _ = network;
+        loop {
+            match event_queue.receive() {
+                NetEvent::Message(_, message) => {
+                    assert_eq!(message.len(), MESSAGE_SIZE);
+                    assert!(message.iter().all(|&byte| byte == VALUE));
+                    break
+                }
+                NetEvent::AddedEndpoint(_) => (),
+                NetEvent::RemovedEndpoint(_) => (),
+            }
+        }
+    });
+
+    let mut event_queue = EventQueue::<NetEvent<Vec<u8>>>::new();
+    let sender = event_queue.sender().clone();
+    let mut network = NetworkManager::new(move |net_event| sender.send(net_event));
+
+    let receiver = network.connect_udp(receiver_addr).unwrap();
+    let message = std::iter::repeat(VALUE).take(MESSAGE_SIZE).collect::<Vec<_>>();
+    network.send(receiver, message).unwrap(); // Blocks until the message is sent
+
+    receiver_handle.join().unwrap();
 }